Move optimizer dump helper to optimizer utils
